+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - 集成 Swoole 服务组件
 - 增加 Twig 扩展, `FastD\Framework\Extensions\TplExtension`
 - 增加注释路由定义方式
+- 调整 FastD\Framework\Bundle\Events\Http\Event -> FastD\Framework\Bundle\Controllers\Controller
 
 
 ### v1.0.* (07/09/2015)